* v1.2.0 - see CHANGELOG for details

+++ template/test/index.test.ts
@@ -1,29 +1,26 @@
-import { expect, test, beforeAll, afterAll } from "bun:test";
-const { spawn } = require('node:child_process');
+/**
+ * Forewords:
+ * This file is just an example of how we may implement test logic for
+ * this project. You can throw this whole file away and write your own
+ * tests using whichever test framework you prefer.
+ */
 
-let cp;
+import { expect, test, beforeAll, afterAll } from "bun:test";
+import { spawn, type ChildProcess } from "node:child_process";
+
+let cp: ChildProcess;
+let cpExited: boolean;
 
 beforeAll(() => {
-  cp = spawn('bun', ['src/index.ts']);
-
-  cp.stdout.on('data', (data) => {
-    console.log(`stdout: ${data}`);
-  });
-
-  cp.stderr.on('data', (data) => {
-    console.error(`stderr: ${data}`);
-  });
-
-  cp.on('close', (code) => {
-    console.log(`test process exited with code ${code}`);
-  });
+  cp = spawn("bun", ["src/index.ts"]);
+  cp.stderr?.on("data", (data) => console.error(`test process stderr: ${data}`));
+  cp.on("error", (err) => console.error(err.stack));
+  cp.on("close", (code) => { cpExited = true; });
 });
 
-const waitFor = (ms = 3000) => new Promise((resolve) => {
-  setTimeout(resolve, ms);
-});
+const waitFor = (ms = 3000) => new Promise((resolve) => setTimeout(resolve, ms));
 
-const tryFetch = async (url, retry = 0, maxRetries = 5) => {
+const tryFetch = async (url: string, retry = 0, maxRetries = 5): Promise<Response> => {
   try {
     return await fetch(url);
   } catch (e) {
@@ -35,29 +32,27 @@ const tryFetch = async (url, retry = 0, maxRetries = 5) => {
 };
 
 function killCp() {
-  if (cp) {
-    console.log(`attempting to kill test process ${cp.pid}`);
-    const killed = cp.kill('SIGKILL');
-    if (!killed) console.error(`test process ${cp.pid} did not exit cleanly, please kill it manually`);
+  if (cp && !cpExited) {
+    const killed = cp.kill("SIGKILL");
+    if (!killed) console.warn(`test process ${cp.pid} may not have exited cleanly, please ensure it is killed`);
   }
 }
 
-process.on('uncaughtException', () => {
-  console.log('bun test uncaughtException');
+process.on("uncaughtException", (e) => {
+  console.error("test: uncaughtException", e.stack);
   killCp();
 });
 
-process.on('beforeExit', () => {
-  console.log('bun test beforeExit');
-  killCp();
-});
+afterAll(() => { killCp(); });
 
-afterAll(() => {
-  killCp();
-});
-
-test('calling /v1/hello/:name', async () => {
-  const response = await tryFetch('http://127.0.0.1:1993/v1/hello/tester');
+test("call /v1/hello/:name", async () => {
+  const response = await tryFetch("http://127.0.0.1:1993/v1/hello/tester");
   const respText = await response.text();
-  expect(respText).toBe('hello, tester');
+  expect(respText).toBe("hello, tester");
+});
+
+test("call /oas.json", async () => {
+  const response = await tryFetch("http://127.0.0.1:1993/oas.json");
+  const respJson = await response.json();
+  expect(respJson).toMatchSnapshot("OpenApiSpecJsonDoc");
 });
